Add exception info to logging message when executing the task fails.
This will give people (me!) a better chance of working out what's gone wrong.

Someone will probably argue that the "correct" way to do this is to use `logging.exception`, but having the traceback for the code in the emulator is unlikely to be particularly helpful. What *is* helpful is seeing the type of error and the exception message, so that you can see what type of error it is (e.g. connection error because you messed up your host/post settings, or an error happening on the server being connected to).